Needs further debugging.
The script can now read the image data correctly...sort of. It places pixels in the correct place in the image, but usually doesn't finish reading all the pixels. You end up with half an image, padded with zeros where it didn't finish reading.

Sometimes it also fails and returns a NULL pointer (see SDK p. 20)